u64

+++ src/felix86/hle/signals.cpp
@@ -573,7 +573,7 @@ void signal_handler(int sig, siginfo_t* info, void* ctx) {
 
         if (jit_code) {
             // If in jit code, make it jump to the dispatcher immediately. If it's not in jit code, just let it naturally go to the dispatcher.
-            context->uc_mcontext.__gregs[REG_PC] = g_emulator->GetRecompiler().getCompileNext();
+            context->uc_mcontext.__gregs[REG_PC] = (u64)g_emulator->GetRecompiler().getCompileNext();
         }
         break;
     }
